Correct spelling and issues introduced by spreadsheet editor.
Change Glutamine Synthetase to Glutamine synthetase to be consistent
with the rest of the data which didn't use upper case for the second word.

Spreadsheet editor changes strings that look like scientific notation
to equivalent numbers, e.g. 9E9 to 9.00E+09, this is a
mistake. Changing to original string values.