Add grants to .zenodo.json description.
Grants in the .zenodo.json are expected to be listed in the grants
list, along the lines of:
"grants": [{"id":"10.13039/100000002::1ZIAAI001290-02"},
           {"id":"10.13039/100000002::1ZIAAI000545-33"}
          ]
Zenodo verifies the grant numbers against their internal
database. This is obtained from OpenAIRE (https://www.openaire.eu/).
Unfortunately, the OpenAIRE database is not up to date and does not
list all grants. This causes Zendo to fail to update as the grant
numbers are not found. The current workaround is to list the grants in
the description text (see https://github.com/zenodo/zenodo/issues/2371).