Modifying the workflow to match the refactored directory structure.
Modified workflow to match directory structure. Also, workflow is now
using a much finer grained control on what to run. The specific
running of scripts was not separated into different workflows because
we want all these changes to go in as one commit when data is
updated. Splitting them up would mean multiple commits and require
synchronization because by default the workflows run in parallel.