Adding an image_resources.csv to the data lake.
Improving the organization of the image data before the data lake
becomes a swamp. The file lists all the image files, their md5 hashes
so that we ensure there is a single copy of each image no matter the
file name, the image caption and which supporting material markdown
files refer to the image.