Correct md5 hashes in image_resources.csv.
Previously md5 hashes were computed for the image in memory using
SimpleITK. We now record the md5 hash of the file on disk.

The usage of the in memory hash was done using SimpleITK and is more
flexible in terms of image content. Thus, if a.jpg and b.png
represent the same image content, it doesn't matter that the files on
disk have different hashes. From a practical standpoint, it is easier
for people to obtain the hash for a file on disk, so switching to that.