Adding shortcut for potential contributors not familiar with GitHub.
Provide an additional data contribution path for researchers who
only provide content occasionally and are not familiar with
GitHub. They can download the latest KB version as a zip file, update
the information there and then contact us for instructions (sharing
the content back with us and then we incorporate it into the KB).